Update to windows1803fs-release
We also want to start moving towards being agnostic of the version. We
started doing this by making the "embed" folder not have the version
inside of it, so instead of updating `/embed/windows2016fs-release` to
`/embed/windows1803fs-release`, we are always embedding the release at
`/embed/windowsfs-release`.

There are still dependencies on the version, but we want to address
these incrementally

[#159601804] [XTP] Garden-windows team wants PASW 2016 2.3 Updates

Signed-off-by: Chris Hendrix <chendrix@pivotal.io>