vendors in the createRelease code
One of the problems we had was that we depended on an external package
called create/createRelease which was provided by the windowsfs-release
that was used as a GOPATH when building the winfs-injector.

I just took all the code that was provided by that release and put it at
the end of main.go. You'll notice that it's still dependent on a blobs
directory for a particularly named release. **This** is one of the many
places that inject a dependency on the particular windows version into
the winfs-injector. This will be made agnostic later.

First I wanted to remove the need to build the injector inside of a
particular windowsfs-release.

[#159676912] The winfs injector CLI works against WRT-2016 2.1, 2.2, and 2.3